# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cloud Photo Delivery API
  version: 0.12.0
  description: API-first backend for the Cloud Photo Delivery SaaS.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      operationId: signup
      summary: Register a new operator account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "423":
          description: Account locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refresh
      summary: Exchange a refresh token for a new token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "401":
          description: Invalid, expired, or reused refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Revoke a refresh token
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /auth/password/reset-request:
    post:
      tags: [Auth]
      operationId: requestPasswordReset
      summary: Request a password reset email (always returns 200)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /auth/password/reset-confirm:
    post:
      tags: [Auth]
      operationId: confirmPasswordReset
      summary: Set a new password using a reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /account/me:
    get:
      tags: [Account]
      operationId: getMe
      summary: Get the authenticated user's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileEnvelope"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    patch:
      tags: [Account]
      operationId: updateMe
      summary: Update the authenticated user's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                businessName:
                  type: string
                  maxLength: 255
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileEnvelope"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /account/branding:
    get:
      tags: [Account]
      operationId: getBranding
      summary: Get the authenticated operator's branding
      description: |
        Returns the tenant's branding. When no branding has been configured the
        string fields are null. Asset keys are replaced by short-lived signed
        URLs. Empty string inputs clear a field on update.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Branding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandingEnvelope"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    patch:
      tags: [Account]
      operationId: updateBranding
      summary: Update the authenticated operator's branding
      description: |
        Partial update: omitted fields are left unchanged; sending an empty
        string clears the field. `logoKey`/`profileImageKey` must come from
        `POST /account/branding/assets`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandingUpdateRequest"
      responses:
        "200":
          description: Updated branding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandingEnvelope"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "402":
          description: Plan does not include custom branding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Validation error (invalid color, email, URL, or asset key)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /account/branding/assets:
    post:
      tags: [Account]
      operationId: createBrandingAssetUpload
      summary: Presign an upload for a branding asset
      description: |
        Returns a presigned PUT URL and the object key for a logo or profile
        image. Upload the bytes straight to storage, then set the key via
        `PATCH /account/branding`. Allowed types are PNG, JPEG, and WebP.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandingAssetRequest"
      responses:
        "201":
          description: Presigned upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandingAssetEnvelope"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "402":
          description: Plan does not include custom branding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Unsupported asset kind or content type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /account/analytics:
    get:
      tags: [Analytics]
      operationId: getAccountAnalytics
      summary: Get analytics aggregated across the operator's events
      description: |
        Returns all-time totals plus a per-day series for the requested window
        (default 30 days, UTC days). Counters are recorded when the public
        gallery is viewed; downloads when an original file is fetched; QR scans
        when the gallery opens with `?src=qr`. `uniqueVisitors` is the sum of
        exact daily uniques, so a visitor returning on another day counts
        again.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AnalyticsDays"
      responses:
        "200":
          description: Account analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountAnalyticsEnvelope"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /events:
    post:
      tags: [Events]
      operationId: createEvent
      summary: Create an event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithSettingsEnvelope"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "402":
          description: Live event limit reached (upcoming, active, and completed count)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    get:
      tags: [Events]
      operationId: listEvents
      summary: List the authenticated operator's events
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/EventStatus"
        - name: q
          in: query
          description: Case-insensitive search on the event name
          schema:
            type: string
        - name: cursor
          in: query
          description: Opaque keyset cursor returned as nextCursor
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListEnvelope"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Invalid status or cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /events/{eventID}:
    parameters:
      - $ref: "#/components/parameters/EventID"
    get:
      tags: [Events]
      operationId: getEvent
      summary: Get one event
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventEnvelope"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "404":
          description: Event not found (also for other tenants)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    patch:
      tags: [Events]
      operationId: updateEvent
      summary: Update event fields
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventEnvelope"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "402":
          description: Expiry beyond the plan's retention window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "409":
          description: The event is expired or archived, so its expiry cannot change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    delete:
      tags: [Events]
      operationId: deleteEvent
      summary: Soft delete an event
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
  /events/{eventID}/photos:
    parameters:
      - $ref: "#/components/parameters/EventID"
    get:
      tags: [Photos]
      operationId: listEventPhotos
      summary: List an event's photos for its owner
      description: |
        Cursor-paginated, newest first, includes photos in every processing
        state (unlike the public gallery, which lists READY photos only).
        Accepts an operator JWT or a device key assigned to the event.
      security:
        - bearerAuth: []
        - deviceKey: []
      parameters:
        - name: cursor
          in: query
          description: Opaque keyset cursor returned as nextCursor
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: status
          in: query
          description: Optional processing status filter
          schema:
            $ref: "#/components/schemas/PhotoStatus"
      responses:
        "200":
          description: A page of photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoListEnvelope"
        "404":
          description: Event not found (also for other tenants)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Invalid cursor, limit, or status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
  /uploads/{photoID}:
    parameters:
      - $ref: "#/components/parameters/PhotoID"
    get:
      tags: [Uploads]
      operationId: getUploadStatus
      summary: Poll the status of an upload (for resume/recovery)
      security:
        - bearerAuth: []
        - deviceKey: []
      responses:
        "200":
          description: Upload status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadStatusEnvelope"
        "404":
          description: Photo not found (also for other tenants)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
  /uploads/{photoID}/multipart/abort:
    parameters:
      - $ref: "#/components/parameters/PhotoID"
    post:
      tags: [Uploads]
      operationId: abortMultipartUpload
      summary: Abort a multipart upload
      security:
        - bearerAuth: []
        - deviceKey: []
      responses:
        "204":
          description: Aborted
        "404":
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "409":
          description: Photo is not a multipart upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Device API key issued by `POST /devices` (format `cpd_live_<prefix>_<secret>`).
        `Authorization: Device <key>` is also accepted.
...
  schemas:
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        businessName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    AuthEnvelope:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/AuthResponse"
        error:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/Error"
    ProfileEnvelope:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Profile"
        error:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/Error"
...
    Envelope:
      type: object
      properties:
        data: {}
        error:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/Error"
...
    EventStatus:
      type: string
      enum: [upcoming, active, completed, archived, expired]
...