# cmd/api/main.go
package main

import (
	"context"
	"errors"
	"flag"
	"log/slog"
	"net/http"
	"os"
	"os/signal"
	"syscall"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/imanjofaris/cloud-photo-delivery/internal/admin"
	"github.com/imanjofaris/cloud-photo-delivery/internal/analytics"
	"github.com/imanjofaris/cloud-photo-delivery/internal/auth"
	"github.com/imanjofaris/cloud-photo-delivery/internal/billing"
	"github.com/imanjofaris/cloud-photo-delivery/internal/billing/provider"
	"github.com/imanjofaris/cloud-photo-delivery/internal/devices"
	"github.com/imanjofaris/cloud-photo-delivery/internal/events"
	"github.com/imanjofaris/cloud-photo-delivery/internal/exports"
	"github.com/imanjofaris/cloud-photo-delivery/internal/gallery"
	"github.com/imanjofaris/cloud-photo-delivery/internal/photos"
	"github.com/imanjofaris/cloud-photo-delivery/internal/platform/config"
	"github.com/imanjofaris/cloud-photo-delivery/internal/platform/logging"
	"github.com/imanjofaris/cloud-photo-delivery/internal/platform/metrics"
	"github.com/imanjofaris/cloud-photo-delivery/internal/qr"
	"github.com/imanjofaris/cloud-photo-delivery/internal/uploads"
	"github.com/imanjofaris/cloud-photo-delivery/internal/users"
	"github.com/imanjofaris/cloud-photo-delivery/pkg/database"
	"github.com/imanjofaris/cloud-photo-delivery/pkg/httpx"
	"github.com/imanjofaris/cloud-photo-delivery/pkg/r2"
)

func main() {
	cfgPath := flag.String("config", "", "path to .env file (optional)")
	_ = cfgPath
	flag.Parse()

	cfg, err := config.Load()
	if err != nil {
		// Logger not yet built; use default.
		logging.New("dev", "info").Error("config load failed", "error", err)
		os.Exit(1)
	}

	log := logging.New(cfg.Env, cfg.LogLevel)

	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
	defer stop()

	pool, err := database.NewWithOptions(ctx, cfg.DatabaseURL, cfg.DatabaseOptions(log))
	if err != nil {
		log.Error("database init failed", "error", err)
		os.Exit(1)
	}
	defer pool.Close()

	m := metrics.New()
	if err := metrics.RegisterDBPool(m.Registry(), func() metrics.PoolStats { return pool.Stat() }); err != nil {
		log.Error("metrics registration failed", "error", err)
		os.Exit(1)
	}

	router := NewRouter(cfg, log, pool, m)

	srv := httpx.NewServer(cfg.HTTPAddr, router)

	go func() {
		log.Info("api server listening", "addr", cfg.HTTPAddr, "env", cfg.Env)
		if err := srv.ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
			log.Error("server error", "error", err)
			stop()
		}
	}()

	metricsSrv := httpx.NewInternalServer(cfg.MetricsAddr, m.Handler())
	if cfg.MetricsAddr != "" {
		go func() {
			log.Info("metrics server listening", "addr", cfg.MetricsAddr)
			if err := metricsSrv.ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
				log.Error("metrics server error", "error", err)
			}
		}()
	}

	<-ctx.Done()
	log.Info("shutting down")

	shutdownCtx, cancel := context.WithTimeout(context.Background(), cfg.ShutdownTimeout)
	defer cancel()
	if err := srv.Shutdown(shutdownCtx); err != nil {
		log.Error("graceful shutdown failed", "error", err)
	}
	if err := metricsSrv.Shutdown(shutdownCtx); err != nil {
		log.Error("metrics shutdown failed", "error", err)
	}
}

func NewRouter(cfg config.Config, log *slog.Logger, pool *database.Pool, m *metrics.Metrics) http.Handler {
	r := chi.NewRouter()

	r.Use(httpx.RequestID)
	r.Use(httpx.Recover)
	r.Use(httpx.Logging(log))
	r.Use(httpx.Metrics(m))
	r.Use(httpx.SecurityHeaders(cfg.IsProd()))
	r.Use(httpx.CORS(cfg.CORSAllowedOrigins))
	r.Use(httpx.MaxBytes(cfg.MaxBodyBytes))
	r.Use(httpx.RequestTimeout(cfg.RequestTimeout))

	r.Get("/healthz", func(w http.ResponseWriter, req *http.Request) {
		httpx.Success(w, http.StatusOK, map[string]string{"status": "ok"})
	})

	r.Get("/readyz", func(w http.ResponseWriter, req *http.Request) {
		ctx, cancel := context.WithTimeout(req.Context(), 3*time.Second)
		defer cancel()
		if err := pool.Ping(ctx); err != nil {
			httpx.Fail(w, http.StatusServiceUnavailable, "NOT_READY", "Database is not available")
			return
		}
		httpx.Success(w, http.StatusOK, map[string]string{"status": "ready"})
	})

	// Rate-limit keys trust X-Forwarded-For only from configured proxies.
	clientIP := httpx.NewClientIPResolver(cfg.TrustedProxyCIDRs)

	// Domain wiring.
	userRepo := users.NewRepository(pool.Pool)
	userSvc := users.NewService(userRepo)
	authRepo := auth.NewRepository(pool.Pool)
	tokenSvc := auth.NewTokenService(cfg.JWTSecret, cfg.AccessTokenTTL, cfg.RefreshTokenTTL)
	mailer := &auth.LogMailer{Log: func(msg string, args ...any) { log.Info(msg, args...) }}
	authSvc := auth.NewService(userRepo, authRepo, tokenSvc, mailer, auth.Config{
		AccessTTL:        cfg.AccessTokenTTL,
		RefreshTTL:       cfg.RefreshTokenTTL,
		PasswordResetTTL: cfg.PasswordResetTTL,
		LockoutMaxFailed: cfg.LockoutMaxFailed,
		LockoutDuration:  cfg.LockoutDuration,
		PublicBaseURL:    cfg.PublicBaseURL,
	})
	authHandler := auth.NewHandler(authSvc)
	userHandler := users.NewHandler(userSvc, func(req *http.Request) (string, bool) {
		id, ok := auth.UserID(req.Context())
		if !ok {
			return "", false
		}
		return id.String(), true
	})

	eventRepo := events.NewRepository(pool.Pool)
	billingRepo := billing.NewRepository(pool.Pool)
	billingProvider := provider.NewManual(cfg.BillingWebhookSecret)
	billingSvc := billing.NewService(billingRepo, billingProvider)
	billingHandler := billing.NewHandler(billingSvc, billingProvider, func(req *http.Request) (string, bool) {
		id, ok := auth.UserID(req.Context())
		if !ok {
			return "", false
		}
		return id.String(), true
	})
	entitlements := billing.NewEntitlements(billingRepo)
	eventSvc := events.NewService(eventRepo, entitlements, auth.HashPassword)
	eventHandler := events.NewHandler(eventSvc, func(req *http.Request) (string, bool) {
		id, ok := auth.UserID(req.Context())
		if !ok {
			return "", false
		}
		return id.String(), true
	})
	qrSvc := qr.NewService(eventSvc, cfg.PublicBaseURL)
	qrHandler := qr.NewHandler(qrSvc, func(req *http.Request) (string, bool) {
		id, ok := auth.UserID(req.Context())
		if !ok {
			return "", false
		}
		return id.String(), true
	})

	photoRepo := photos.NewRepository(pool.Pool)
	uploadRepo := uploads.NewRepository(pool.Pool)
	var store r2.ObjectStore = r2.UnavailableStore{}
	if err := cfg.ValidateStorage(); err == nil {
		s, err := r2.New(context.Background(), r2.Options{
			Endpoint:  cfg.R2Endpoint,
			AccessKey: cfg.R2AccessKey,
			SecretKey: cfg.R2SecretKey,
			Bucket:    cfg.R2Bucket,
			Region:    cfg.R2Region,
		})
		if err != nil {
			log.Error("object store init failed", "error", err)
			os.Exit(1)
		}
		store = s
	}
	store = metrics.WrapStore(store, m)
	brandingSvc := users.NewBrandingService(userRepo, store, cfg.SignedURLTTL, entitlements)
	brandingHandler := users.NewBrandingHandler(brandingSvc, func(req *http.Request) (string, bool) {
		id, ok := auth.UserID(req.Context())
		if !ok {
			return "", false
		}
		return id.String(), true
	})
	uploadSvc := uploads.NewService(photoRepo, uploadRepo, store, uploads.NewPostgresQueue(pool.Pool), entitlements).WithObserver(m)
	uploadHandler := uploads.NewHandler(uploadSvc, func(req *http.Request) (uploads.Actor, bool) {
		if device, ok := devices.FromContext(req.Context()); ok {
			return uploads.Actor{
				UserID:        device.UserID,
				DeviceID:      device.ID,
				AssignedEvent: device.AssignedEventID,
			}, true
		}
		id, ok := auth.UserID(req.Context())
		if !ok {
			return uploads.Actor{}, false
		}
		return uploads.Actor{UserID: id}, true
	})

	deviceRepo := devices.NewRepository(pool.Pool)
	deviceSvc := devices.NewService(deviceRepo, entitlements)
	deviceHandler := devices.NewHandler(deviceSvc, func(req *http.Request) (string, bool) {
		id, ok := auth.UserID(req.Context())
		if !ok {
			return "", false
		}
		return id.String(), true
	})

	analyticsSvc := analytics.NewService(
		analytics.NewRepository(pool.Pool), photoRepo, cfg.AnalyticsSalt, log)
	analyticsHandler := analytics.NewHandler(analyticsSvc, func(req *http.Request) (string, bool) {
		id, ok := auth.UserID(req.Context())
		if !ok {
			return "", false
		}
		return id.String(), true
	})

	adminSvc := admin.NewService(admin.NewRepository(pool.Pool))
	adminHandler := admin.NewHandler(adminSvc)

	galleryRepo := gallery.NewRepository(pool.Pool)
	galleryURLs := photos.NewSignedURLGenerator(store, cfg.SignedURLTTL)
	galleryTokens := gallery.NewUnlockTokens(cfg.JWTSecret, cfg.GalleryUnlockTTL)
	gallerySvc := gallery.NewService(galleryRepo, galleryURLs, galleryTokens, auth.VerifyPassword, brandingSvc, analyticsSvc, entitlements)
	galleryHandler := gallery.NewHandler(gallerySvc, clientIP.Resolve)

	photoURLs := photos.NewSignedURLGenerator(store, cfg.SignedURLTTL)
	photoSvc := photos.NewService(photoRepo, photoURLs, store, photos.NewPostgresQueue(pool.Pool))
	photoHandler := photos.NewHandler(photoSvc, func(req *http.Request) (photos.Actor, bool) {
		if device, ok := devices.FromContext(req.Context()); ok {
			return photos.Actor{
				UserID:        device.UserID,
				DeviceID:      device.ID,
				AssignedEvent: device.AssignedEventID,
			}, true
		}
		id, ok := auth.UserID(req.Context())
		if !ok {
			return photos.Actor{}, false
		}
		return photos.Actor{UserID: id}, true
	})

	exportSvc := exports.NewService(
		exports.NewRepository(pool.Pool), photoRepo, photoRepo,
		exports.NewPostgresQueue(pool.Pool), store, cfg.ExportTTL)
	exportHandler := exports.NewHandler(exportSvc, func(req *http.Request) (string, bool) {
		id, ok := auth.UserID(req.Context())
		if !ok {
			return "", false
		}
		return id.String(), true
	})

	// Phase 10 starting values (doc/phases/Phase 10 - Hardening and Deploy.md
	// §2), with signed URL generation re-sized for batched gallery reads: the
	// limiter charges one token per URL signed, not per request, so a
	// 120-photo page needs ~120 tokens instead of 120 requests.
	loginLimiter := httpx.NewRateLimiter(10, 10, 10*time.Minute)
	signupLimiter := httpx.NewRateLimiter(5, 5, 10*time.Minute)
	adminLimiter := httpx.NewRateLimiter(30, 30, 10*time.Minute)
	uploadInitLimiter := httpx.NewRateLimiter(100, 100, 10*time.Minute)
	uploadCompleteLimiter := httpx.NewRateLimiter(300, 300, 10*time.Minute)
	signedURLLimiter := httpx.NewRateLimiter(600, 300, 10*time.Minute)

	userKey := func(req *http.Request) string {
		id, ok := auth.UserID(req.Context())
		if !ok {
			return "anonymous"
		}
		return "user:" + id.String()
	}
	uploadKey := func(req *http.Request) string {
		if device, ok := devices.FromContext(req.Context()); ok {
			return "device:" + device.ID.String()
		}
		return userKey(req)
	}

	r.Route("/api/v1", func(r chi.Router) {
		r.With(signupLimiter.Middleware(clientIP.Resolve)).Post("/auth/signup", authHandler.Signup)

		r.Group(func(r chi.Router) {
			r.Use(loginLimiter.Middleware(clientIP.Resolve))
			r.Post("/auth/login", authHandler.Login)
			r.Post("/auth/refresh", authHandler.Refresh)
			r.Post("/auth/logout", authHandler.Logout)
			r.Post("/auth/password/reset-request", authHandler.RequestPasswordReset)
			r.Post("/auth/password/reset-confirm", authHandler.ConfirmPasswordReset)
		})

		// Public gallery (no auth). Signed URL generation is rate limited by
		// client IP per event, charging one token per URL signed so batched
		// requests share the same budget as single-photo requests.
		r.Route("/public/events/{slug}", func(r chi.Router) {
			r.Get("/", galleryHandler.GetEvent)
			r.Post("/unlock", galleryHandler.Unlock)
			r.Get("/photos", galleryHandler.ListPhotos)
			r.With(signedURLLimiter.MiddlewareN(gallery.SignedURLRateLimitKey(clientIP.Resolve), gallery.BatchURLCost)).
				Post("/photos/urls", galleryHandler.BatchPhotoURLs)
			r.Get("/photos/{photoID}", galleryHandler.GetPhoto)
			r.With(signedURLLimiter.MiddlewareN(gallery.SignedURLRateLimitKey(clientIP.Resolve), nil)).
				Get("/photos/{photoID}/url", galleryHandler.PhotoURL)
		})

		// Provider callback: unauthenticated but signature-verified.
		r.Post("/billing/webhook", billingHandler.Webhook)

		r.Group(func(r chi.Router) {
			r.Use(authSvc.RequireAuth)
			r.Get("/account/me", userHandler.Me)
			r.Patch("/account/me", userHandler.UpdateMe)
			r.Get("/account/branding", brandingHandler.Get)
			r.Patch("/account/branding", brandingHandler.Update)
			r.Post("/account/branding/assets", brandingHandler.CreateAssetUpload)
			r.Get("/account/analytics", analyticsHandler.Account)

			r.Route("/billing", func(r chi.Router) {
				r.Get("/plans", billingHandler.Plans)
				r.Get("/subscription", billingHandler.GetSubscription)
				r.Post("/subscribe", billingHandler.Subscribe)
				r.Post("/upgrade", billingHandler.Upgrade)
				r.Post("/downgrade", billingHandler.Downgrade)
				r.Post("/cancel", billingHandler.Cancel)
				r.Post("/resume", billingHandler.Resume)
				r.Get("/invoices", billingHandler.Invoices)
			})

			r.Route("/events", func(r chi.Router) {
				r.Post("/", eventHandler.Create)
				r.Get("/", eventHandler.List)

				r.Route("/{eventID}", func(r chi.Router) {
					r.Get("/", eventHandler.Get)
					r.Patch("/", eventHandler.Update)
					r.Post("/archive", eventHandler.Archive)
					r.Post("/status", eventHandler.UpdateStatus)
					r.Post("/extend", eventHandler.Extend)
					r.Delete("/", eventHandler.Delete)

					r.Get("/settings", eventHandler.GetSettings)
					r.Patch("/settings", eventHandler.UpdateSettings)
					r.Post("/exports", exportHandler.Create)
					r.Get("/exports/{exportID}", exportHandler.Get)
					r.Get("/analytics", analyticsHandler.Event)
					r.Get("/dashboard", eventHandler.Dashboard)
					r.Get("/url", qrHandler.URL)
					r.Get("/qr.png", qrHandler.PNG)
					r.Get("/qr.svg", qrHandler.SVG)
				})
			})

			r.Route("/devices", func(r chi.Router) {
				r.Post("/", deviceHandler.Create)
				r.Get("/", deviceHandler.List)
				r.Route("/{deviceID}", func(r chi.Router) {
					r.Patch("/", deviceHandler.Update)
					r.Delete("/", deviceHandler.Revoke)
					r.Post("/rotate", deviceHandler.Rotate)
				})
			})
		})

		// Platform administration. RequireAdmin runs after RequireAuth and
		// rejects authenticated non-admin accounts.
		r.Group(func(r chi.Router) {
			r.Use(authSvc.RequireAuth, adminSvc.RequireAdmin, adminLimiter.Middleware(userKey))
			r.Get("/admin/stats", adminHandler.Stats)
			r.Get("/admin/users", adminHandler.Users)
			r.Get("/admin/subscriptions", adminHandler.Subscriptions)
			r.Get("/admin/health", adminHandler.Health)
		})

		// Uploads accept an operator JWT or a device key scoped to the event.
		r.Group(func(r chi.Router) {
			r.Use(deviceSvc.RequireDeviceOrOperator(authSvc.RequireAuth))
			r.With(uploadInitLimiter.Middleware(uploadKey)).
				Post("/events/{eventID}/uploads", uploadHandler.Initialize)
			r.Get("/events/{eventID}/photos", photoHandler.List)
			r.Route("/uploads/{photoID}", func(r chi.Router) {
				r.Get("/", uploadHandler.Status)
				r.Post("/url", uploadHandler.RePresign)
				r.Post("/parts", uploadHandler.Parts)
				r.With(uploadCompleteLimiter.Middleware(userKey)).
					Post("/multipart/complete", uploadHandler.CompleteMultipart)
				r.Post("/multipart/abort", uploadHandler.AbortMultipart)
				r.With(uploadCompleteLimiter.Middleware(userKey)).
					Post("/complete", uploadHandler.Complete)
			})
			r.Route("/photos/{photoID}", func(r chi.Router) {
				r.With(signedURLLimiter.Middleware(clientIP.Resolve)).Get("/url", photoHandler.URL)
				r.Delete("/", photoHandler.Delete)
			})
		})
	})

	return r
}